refactor(web2): extract base class ItemFromBase
