+++ web2/src/components/InventoryForm.vue
@@ -29,99 +29,17 @@
   </v-form>
 </template>
 <script>
-import axios from 'axios';
-import { getErrorMessage } from '@/lib/error';
+import ItemFormBase from '@/components/ItemFormBase';
 
 export default {
-  props: {
-    itemId: [Number, String],
-    projectId: [Number, String],
-    needSave: Boolean,
-    needReset: Boolean,
-  },
-
-  data() {
-    return {
-      item: null,
-      formValid: false,
-      formError: null,
-      formSaving: false,
-    };
-  },
-
-  async created() {
-    await this.loadData();
-  },
-
-  computed: {
-    isNew() {
-      return this.itemId === 'new';
-    },
-  },
-
-  watch: {
-    needSave(val) {
-      if (val) {
-        this.save();
-      }
-    },
-    needReset(val) {
-      if (val) {
-        this.reset();
-      }
-    },
-  },
-
+  mixins: [ItemFormBase],
   methods: {
-    async reset() {
-      this.item = null;
-      this.$refs.form.resetValidation();
-      await this.loadData();
+    getItemsUrl() {
+      return `/api/project/${this.projectId}/inventory`;
     },
 
-    async loadData() {
-      if (this.isNew) {
-        this.item = {};
-      } else {
-        this.item = (await axios({
-          method: 'get',
-          url: `/api/project/${this.projectId}/inventory/${this.itemId}`,
-          responseType: 'json',
-        })).data;
-      }
-    },
-
-    /**
-     * Saves or creates user via API.
-     * @returns {Promise<null>} null if validation didn't pass or user data if user saved.
-     */
-    async save() {
-      this.formError = null;
-
-      if (!this.$refs.form.validate()) {
-        return;
-      }
-
-      this.formSaving = true;
-      try {
-        const item = (await axios({
-          method: this.isNew ? 'post' : 'put',
-          url: this.isNew
-            ? `/api/project/${this.projectId}/inventory`
-            : `/api/project/${this.projectId}/inventory/${this.item.id}`,
-          responseType: 'json',
-          data: this.item,
-        })).data;
-
-        this.$emit('save', {
-          item,
-          action: this.isNew ? 'new' : 'edit',
-        });
-      } catch (err) {
-        this.formError = getErrorMessage(err);
-      } finally {
-        this.formSaving = false;
-      }
+    getSingleItemUrl() {
+      return `/api/project/${this.projectId}/inventory/${this.itemId}`;
     },
   },
 };
